refactor:move product-related functionality from script.js to products.page
