+++ script.js
@@ -1,23 +1,12 @@
 import { ProductsPage } from "./src/pages/products.page.js";
 import { CartPage } from "./src/pages/cart.page.js";
-import {
-  getProductById,
-  setCategory,
-  setMaxPrice,
-  setSearchText,
-  setSortText,
-} from "./src/features/products/product.store.js";
+import { getProductById } from "./src/features/products/product.store.js";
 import {
   addToCart,
   decreaseQty,
   toggleCart,
 } from "./src/features/cart/cart.store.js";
 import { showToast } from "./utils.js";
-
-const dropdown = document.getElementById("categoryDropdown");
-const trigger = dropdown.querySelector(".dropdown__trigger");
-const menu = dropdown.querySelector(".dropdown__menu");
-const selected = document.getElementById("selectedCategory");
 
 document.addEventListener("DOMContentLoaded", () => {
   ProductsPage();
@@ -44,39 +33,6 @@ document
   .getElementById("closeCart")
   .addEventListener("click", handleCartToggle);
 document.getElementById("backdrop").addEventListener("click", handleCartToggle);
-
-document.getElementById("priceRange").addEventListener("input", (e) => {
-  setMaxPrice(Number(e.target.value));
-  document.getElementById("priceLabel").textContent = e.target.value;
-  ProductsPage();
-});
-
-document.addEventListener("click", (e) => {
-  if (e.target.classList.contains("category-list__btn")) {
-    const category = e.target.dataset.category;
-    setCategory(category);
-    ProductsPage();
-  }
-});
-
-document.getElementById("searchInput").addEventListener("input", (e) => {
-  setSearchText(e.target.value);
-  ProductsPage();
-});
-
-// open/close dropdown
-trigger.addEventListener("click", () => {
-  dropdown.classList.toggle("open");
-});
-
-menu.addEventListener("click", (e) => {
-  const value = e.target.dataset.value;
-  if (!value) return;
-  setSortText(value);
-  selected.textContent = e.target.textContent;
-  dropdown.classList.remove("open");
-  ProductsPage();
-});
 
 document.querySelector(".cart-overlay").addEventListener("click", (e) => {
   const cartItem = e.target.closest(".cart-item");
